extra message fetcher out as helper function
